backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
celery
pandas
numpy
scikit-learn
xgboost
joblib
python-dotenv
PyJWT
bcrypt